### 2017-07-21 #324 @whedon
[PRE REVIEW]: rtimicropem: an R package supporting the analysis of RTI MicroPEM output files
**Submitting author:** @maelle (<a href="http://orcid.org/0000-0002-2815-0399">Maëlle Salmon</a>)
**Repository:** <a href="https://github.com/ropensci/rtimicropem" target ="_blank">https://github.com/ropensci/rtimicropem</a>
**Version:** v1.3
