### 2017-07-25 #324 @maelle
Link to rOpenSci review of `rtimicropem` https://github.com/ropensci/onboarding/issues/126
