### 2017-07-26 #333 @arfon
rOpenSci onboarding here https://github.com/ropensci/onboarding/issues/126
