### 2017-07-26 #333 @arfon
@maelle - could you please make sure you've cited the reference in your `paper.bib` file? Otherwise it won't be compiled in when we finish up the submission.

In addition, could you please expand a little on the `paper.md` file? We (the JOSS editors) are in the process of adding more guidance for authors, including a better indication of expected paper length. Currently it's hard to get a picture of the functionality of the software, and also include a statement of need, in less than 250 words. Would you be able to put a little more detail in your submission please?
