### 2017-07-26 #333 @arfon
Also, I've made a couple of small fixes here https://github.com/ropensci/rtimicropem/pull/31
