### 2017-07-27 #333 @maelle
@arfon thanks for handling this submission. I've added the reference in the paper, and added details in `paper.md`.
