### 2017-07-27 #333 @arfon
OK thanks. BTW, it looks like you've committed a merge conflict in [`paper.md`](https://github.com/ropensci/rtimicropem/blob/master/paper/paper.md)
