### 2017-07-27 #333 @maelle
@arfon thanks, I've now removed the conflict.
